requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2
pyyaml==6.0.1
pytest==7.4.3
flake8==6.1.0
black==23.12.1